refactor(oleada-3): extract error-recovery + post-loop drivers from flow-runner
First milestone of TSK-0335 (decompose flow-runner.js, KJC-PCS-0040). Two of
the planned four drivers land in this commit:

**src/orchestrator/drivers/error-recovery.js** (283 LOC)
- handleStandbyResult — rate-limit / provider-outage delegation to Solomon
- emitSolomonAlerts — forward rule findings to the progress stream
- handleSolomonCheck — Solomon-rules evaluation + Brain/legacy branching
- checkSolomonCriticalAlerts — non-Brain critical-alert prompt
- handleReviewerRetryAndSolomon — reviewer sub-loop limit → Solomon

**src/orchestrator/drivers/post-loop.js** (365 LOC)
- handlePostLoopStages — tester → security → final audit
- finalizeApprovedSession — commit/push/PR + PG + journal + session:end
- handleMaxIterationsReached — Brain-owned or Solomon-driven fallback
- tryAutoStartBoard — HU Board auto-launch
- writeHistoryRecord — non-blocking HU history write

flow-runner.js is now 1688 LOC (down from 2254, -25%). Behavior unchanged —
functions moved verbatim, flow-runner imports them. Each driver file has
single responsibility + its own docstring.

Remaining TSK-0335 work (subsequent PRs):
- Extract pre-loop driver (runPlanningPhases, runPreLoopStages,
  emitConfigDeprecations, ensureAddyosmaniSkills, maybeGenerateAutoHuBatch,
  initFlowContext — ~760 LOC, may split into pre-loop.js + init-context.js
  to keep under the 600-LOC/driver ceiling).
- Extract iteration-loop driver (runCoderAndRefactorerStages, runGuardStages,
  runQualityGateStages, runReviewerGateStage, handleApprovedReview,
  runSingleIteration, runIterationLoop).
- After both: flow-runner.js keeps only runFlow + resumeFlow, wiring the
  drivers.

All 3744 tests green.
