+++ src/orchestrator/drivers/post-loop.js
@@ -1,0 +1,365 @@
+/**
+ * Post-loop driver — extracted from src/orchestrator/flow-runner.js in
+ * TSK-0335 (Oleada 3 of the v2.7.4 audit refactor).
+ *
+ * Handlers for everything that happens AFTER the iteration loop decides
+ * an outcome (approved, max-iterations, or simple finalization):
+ *
+ *   - handlePostLoopStages:        run tester → security → final audit
+ *                                  after the reviewer has approved; any of
+ *                                  them can demand one more coder pass.
+ *   - finalizeApprovedSession:     commit/push/PR + PG card transition +
+ *                                  full session journal + session:end event.
+ *   - handleMaxIterationsReached:  Brain-owned decision at the iteration
+ *                                  cap (extend vs finalize vs escalate).
+ *                                  Legacy path delegates to Solomon.
+ *   - tryAutoStartBoard:           auto-launch HU Board when config opts in.
+ *   - writeHistoryRecord:          non-blocking HU history write on session end.
+ *
+ * Extracted verbatim; no behavior change. flow-runner.js now imports these
+ * instead of defining them. Part of breaking up the 2254-LOC god-module —
+ * see docs/ARCHITECTURE.md and the TSK-0335 acceptance criteria.
+ */
+
+import { generateDiff } from "../../review/diff-generator.js";
+import { finalizeGitAutomation } from "../../git/automation.js";
+import { saveSession, markSessionStatus } from "../../session/store.js";
+import { emitProgress, makeEvent } from "../../utils/events.js";
+import { runTesterStage, runSecurityStage, runFinalAuditStage } from "../post-loop-stages.js";
+import { invokeSolomon } from "../solomon-escalation.js";
+import { tryCiComment } from "../ci-integration.js";
+import { writeIterationsJournal } from "../session-journal.js";
+
+export async function handlePostLoopStages({ config, session, emitter, eventBase, coderRole, trackBudget, i, task, stageResults, ciEnabled, testerEnabled, securityEnabled, askQuestion, logger, brainCtx }) {
+  const postLoopDiff = await generateDiff({ baseRef: session.session_start_sha });
+
+  if (testerEnabled) {
+    const testerResult = await runTesterStage({
+      config, logger, emitter, eventBase, session, coderRole, trackBudget,
+      iteration: i, task, diff: postLoopDiff, askQuestion
+    });
+    if (testerResult.action === "pause") return { action: "return", result: testerResult.result };
+    if (testerResult.action === "continue") {
+      const summary = testerResult.stageResult?.summary || "Tester found issues";
+      session.last_reviewer_feedback = `Tester FAILED — fix these issues:\n${summary}`;
+      await saveSession(session);
+      if (testerResult.stageResult) stageResults.tester = testerResult.stageResult;
+      // Brain: push tester failure into feedback queue + compress for next coder iteration
+      if (brainCtx?.enabled) {
+        const { processRoleOutput } = await import("../brain-coordinator.js");
+        processRoleOutput(brainCtx, { roleName: "tester", output: testerResult.stageResult, iteration: i });
+      }
+      return { action: "continue" };
+    }
+    if (testerResult.stageResult) {
+      stageResults.tester = testerResult.stageResult;
+      await tryCiComment({ config, session, logger, agent: "Tester", body: `Tests: ${testerResult.stageResult.summary || "completed"}` });
+    }
+  }
+
+  if (securityEnabled) {
+    const securityResult = await runSecurityStage({
+      config, logger, emitter, eventBase, session, coderRole, trackBudget,
+      iteration: i, task, diff: postLoopDiff, askQuestion, brainCtx
+    });
+    if (securityResult.action === "pause") return { action: "return", result: securityResult.result };
+    if (securityResult.action === "continue") {
+      const summary = securityResult.stageResult?.summary || "Security found issues";
+      session.last_reviewer_feedback = `Security FAILED — fix these issues:\n${summary}`;
+      await saveSession(session);
+      if (securityResult.stageResult) stageResults.security = securityResult.stageResult;
+      // Brain: push security findings into feedback queue + compress for next coder iteration
+      if (brainCtx?.enabled) {
+        const { processRoleOutput } = await import("../brain-coordinator.js");
+        processRoleOutput(brainCtx, { roleName: "security", output: securityResult.stageResult, iteration: i });
+      }
+      return { action: "continue" };
+    }
+    if (securityResult.stageResult) {
+      stageResults.security = securityResult.stageResult;
+      await tryCiComment({ config, session, logger, agent: "Security", body: `Security scan: ${securityResult.stageResult.summary || "completed"}` });
+    }
+  }
+
+  // Final audit — last quality gate before declaring success
+  const auditResult = await runFinalAuditStage({
+    config, logger, emitter, eventBase, session, coderRole, trackBudget,
+    iteration: i, task, diff: postLoopDiff
+  });
+  if (auditResult.stageResult) {
+    stageResults.audit = auditResult.stageResult;
+    await tryCiComment({ config, session, logger, agent: "Audit", body: `Final audit: ${auditResult.stageResult.summary || "completed"}` });
+  }
+  if (auditResult.action === "retry") {
+    // Audit found actionable issues — loop back to coder
+    session.last_reviewer_feedback = auditResult.feedback;
+    await saveSession(session);
+    return { action: "continue" };
+  }
+
+  return { action: "proceed" };
+}
+
+export async function finalizeApprovedSession({ config, gitCtx, task, logger, session, stageResults, emitter, eventBase, budgetSummary, pgCard, pgProject, review, i, rtkTracker }) {
+  const gitResult = await finalizeGitAutomation({ config, gitCtx, task, logger, session, stageResults });
+
+  // Accumulate final commits for PG card lifecycle tracking
+  if (gitResult?.commits?.length) {
+    const { accumulateCommit } = await import("../../planning-game/pipeline-adapter.js");
+    for (const c of gitResult.commits) accumulateCommit(session, c);
+  }
+
+  if (stageResults.planner?.ok) {
+    stageResults.planner.completedSteps = [...(stageResults.planner.steps ?? [])];
+  }
+  session.budget = budgetSummary();
+  await markSessionStatus(session, "approved");
+
+  const { markPgCardToValidate } = await import("../../planning-game/pipeline-adapter.js");
+  await markPgCardToValidate({ pgCard, pgProject, config, session, gitResult, logger });
+
+  const deferredIssues = session.deferred_issues || [];
+  const rtkSavings = rtkTracker?.hasData() ? rtkTracker.summary() : undefined;
+  if (rtkSavings) session.rtk_savings = rtkSavings;
+  await saveSession(session);
+
+  // --- Journal: write final files ---
+  const journalDir = session._journalDir;
+  if (journalDir) {
+    try {
+      await writeIterationsJournal(journalDir, session._journalIterations || []);
+      // decisions.md: writes only when Solomon was invoked (AC of KJC-TSK-0287).
+      // Pulls structured data from session.solomonRulings + session.brainDecisions.
+      const { writeDecisionsJournal: writeDecisionsJournalV2 } =
+        await import("../../session/journal/decisions-writer.js");
+      const decisionsResult = await writeDecisionsJournalV2(session, { journalDir, logger });
+      // KJC-TSK-0288: use the new tree-writer with directory grouping
+      const { writeTreeJournal: writeTreeJournalV2 } =
+        await import("../../session/journal/tree-writer.js");
+      const treeResult = await writeTreeJournalV2(journalDir, session.session_start_sha, { logger });
+
+      const journalFiles = [...(session._journalFiles || [])];
+      if (session._journalIterations?.length) journalFiles.push("iterations.md");
+      if (decisionsResult.written) journalFiles.push("decisions.md");
+      if (treeResult.written) journalFiles.push("tree.txt");
+
+      // KJC-TSK-0289: use the richer summary writer (stages table, budget
+      // breakdown, brain/solomon counts, typed links to other journal files).
+      const { writeSummaryJournal: writeSummaryJournalV2 } =
+        await import("../../session/journal/summary-writer.js");
+      const startedAt = session._startedAt ? new Date(session._startedAt).toISOString() : undefined;
+      await writeSummaryJournalV2(journalDir, {
+        task: session.task,
+        result: "APPROVED",
+        sessionId: session.id,
+        iterations: i,
+        durationMs: Date.now() - (session._startedAt || Date.now()),
+        budget: budgetSummary(),
+        stages: stageResults,
+        commits: gitResult?.commits || [],
+        files: journalFiles,
+        startedAt,
+        finishedAt: new Date().toISOString(),
+        brainDecisions: Array.isArray(session.brainDecisions) ? session.brainDecisions.length : 0,
+        solomonInvocations: Array.isArray(session.solomonRulings) ? session.solomonRulings.length : 0,
+        // KJC-TSK-0327: surface skill-source decisions in the summary so the
+        // post-mortem shows which process skills (addyosmani) and stack skills
+        // (OpenSkills) actually shaped the role prompts.
+        skills: {
+          addyosmani: session.addyosmaniSkills,
+          installed: session.autoInstalledSkills,
+          recommended: session.skillsRecommended,
+        },
+      }, { logger });
+      logger.info(`Session journal written to ${journalDir}`);
+    } catch (err) {
+      logger.warn(`Journal write failed (non-blocking): ${err.message}`);
+    }
+  }
+
+  const endDetail = { approved: true, iterations: i, stages: stageResults, git: gitResult, budget: budgetSummary(), deferredIssues };
+  if (rtkSavings) endDetail.rtk_savings = rtkSavings;
+  emitProgress(
+    emitter,
+    makeEvent("session:end", { ...eventBase, stage: "done" }, {
+      message: deferredIssues.length > 0
+        ? `Session approved (${deferredIssues.length} deferred issue(s) tracked as tech debt)`
+        : "Session approved",
+      detail: endDetail
+    })
+  );
+  const result = { approved: true, sessionId: session.id, review, git: gitResult, deferredIssues };
+  if (rtkSavings) result.rtk_savings = rtkSavings;
+  return result;
+}
+
+export async function handleMaxIterationsReached({ session, budgetSummary, emitter, eventBase, config, stageResults, logger, askQuestion, task, rtkTracker, brainCtx }) {
+  const budget = budgetSummary();
+
+  // Brain-owned decision: max_iterations is guidance, not a hard rule.
+  // Brain evaluates the feedback queue state to decide extend / finalize / escalate.
+  // Solomon is only consulted if Brain cannot decide on its own.
+  if (brainCtx?.enabled) {
+    const entries = brainCtx.feedbackQueue?.entries || [];
+    const pending = entries.map(e => ({ source: e.source, category: e.category, severity: e.severity, description: e.description }));
+    const hasSecurity = entries.some(e => e.category === "security" || e.source === "security");
+    const hasCorrectness = entries.some(e => ["correctness", "tests"].includes(e.category));
+    const hasStyleOnly = entries.length > 0 && !hasSecurity && !hasCorrectness;
+
+    if (hasSecurity) {
+      // Brain: security issues unresolved → cannot finalize, escalate
+      logger.warn(`Brain: max_iterations reached with ${entries.filter(e => e.category === "security" || e.source === "security").length} unresolved security issue(s) — cannot finalize`);
+      return { paused: true, sessionId: session.id, question: "Brain: unresolved security issues at max_iterations. Review manually or extend pipeline.", context: "brain_security_block", pending };
+    }
+
+    if (hasCorrectness) {
+      // Brain: correctness/test issues pending. Cap at MAX_EXTENSIONS to avoid infinite extensions.
+      const MAX_EXTENSIONS = 2;
+      if (brainCtx.extensionCount >= MAX_EXTENSIONS) {
+        logger.warn(`Brain: ${brainCtx.extensionCount} extensions exhausted with correctness issues still pending — escalating to human`);
+        return { paused: true, sessionId: session.id, question: `Brain exhausted ${MAX_EXTENSIONS} extensions with correctness/tests still pending. Manual intervention needed.`, context: "brain_extension_cap", pending };
+      }
+      brainCtx.extensionCount += 1;
+      logger.info(`Brain: max_iterations reached with ${entries.filter(e => ["correctness", "tests"].includes(e.category)).length} correctness issue(s) pending — extending iterations (extension ${brainCtx.extensionCount}/${MAX_EXTENSIONS})`);
+      session.reviewer_retry_count = 0;
+      await saveSession(session);
+      return { approved: false, sessionId: session.id, reason: "max_iterations_extended", extraIterations: Math.ceil(config.max_iterations / 2) };
+    }
+
+    if (entries.length === 0) {
+      // Brain: no pending feedback → last reviewer approved, finalize
+      logger.info("Brain: max_iterations reached with clean feedback queue — finalizing as approved");
+      return { approved: true, sessionId: session.id, reason: "brain_approved" };
+    }
+
+    // hasStyleOnly: genuine dilemma → Brain consults Solomon
+    logger.info(`Brain: max_iterations with ${entries.length} style-only issue(s) — consulting Solomon on dilemma`);
+    const { invokeSolomon: invokeSolomonAI } = await import("../solomon-escalation.js");
+    const solomonResult = await invokeSolomonAI({
+      config, logger, emitter, eventBase, stage: "max_iterations", askQuestion, session,
+      iteration: config.max_iterations,
+      conflict: {
+        stage: "brain-max-iterations",
+        task,
+        iterationCount: config.max_iterations,
+        maxIterations: config.max_iterations,
+        budget_usd: budget?.total_cost_usd || 0,
+        dilemma: `Max iterations reached with ${entries.length} style-only issue(s) pending. Accept as-is or request more work?`,
+        pendingIssues: pending,
+        history: [{ agent: "pipeline", feedback: session.last_reviewer_feedback || "Max iterations reached" }]
+      }
+    });
+    // Brain applies Solomon's decision
+    if (solomonResult.action === "approve") {
+      logger.info("Brain: Solomon advised approve for style-only pending — finalizing");
+      return { approved: true, sessionId: session.id, reason: "brain_solomon_approved" };
+    }
+    if (solomonResult.action === "continue") {
+      return { approved: false, sessionId: session.id, reason: "max_iterations_extended", extraIterations: solomonResult.extraIterations || Math.ceil(config.max_iterations / 2) };
+    }
+    if (solomonResult.action === "pause") {
+      return { paused: true, sessionId: session.id, question: solomonResult.question, context: "brain_solomon_dilemma" };
+    }
+    // Fallback: escalate to human
+    return { paused: true, sessionId: session.id, question: `Brain+Solomon cannot resolve: ${entries.length} pending issue(s) at max_iterations`, context: "max_iterations" };
+  }
+
+  // Legacy path (Brain disabled): original Solomon-driven flow
+  const solomonResult = await invokeSolomon({
+    config, logger, emitter, eventBase, stage: "max_iterations", askQuestion, session,
+    iteration: config.max_iterations,
+    conflict: {
+      stage: "max_iterations",
+      task,
+      iterationCount: config.max_iterations,
+      maxIterations: config.max_iterations,
+      budget_usd: budget?.total_cost_usd || 0,
+      history: [{ agent: "pipeline", feedback: session.last_reviewer_feedback || "Max iterations reached without reviewer approval" }]
+    }
+  });
+
+  if (solomonResult.action === "approve") {
+    logger.info("Solomon approved coder's work at max_iterations checkpoint");
+    return { approved: true, sessionId: session.id, reason: "solomon_approved" };
+  }
+
+  if (solomonResult.action === "continue") {
+    if (solomonResult.humanGuidance) {
+      session.last_reviewer_feedback = `Solomon guidance: ${solomonResult.humanGuidance}`;
+    }
+    session.reviewer_retry_count = 0;
+    await saveSession(session);
+    const extraIterations = solomonResult.extraIterations || config.max_iterations;
+    return { approved: false, sessionId: session.id, reason: "max_iterations_extended", humanGuidance: solomonResult.humanGuidance, extraIterations };
+  }
+
+  if (solomonResult.action === "pause") {
+    return { paused: true, sessionId: session.id, question: solomonResult.question, context: "max_iterations" };
+  }
+
+  if (solomonResult.action === "subtask") {
+    return { paused: true, sessionId: session.id, subtask: solomonResult.subtask, context: "max_iterations_subtask" };
+  }
+
+  // Solomon couldn't resolve — fail
+  session.budget = budgetSummary();
+  const rtkSavings = rtkTracker?.hasData() ? rtkTracker.summary() : undefined;
+  if (rtkSavings) session.rtk_savings = rtkSavings;
+  await markSessionStatus(session, "failed");
+  const failDetail = { approved: false, reason: "max_iterations", iterations: config.max_iterations, stages: stageResults, budget: budgetSummary() };
+  if (rtkSavings) failDetail.rtk_savings = rtkSavings;
+  emitProgress(
+    emitter,
+    makeEvent("session:end", { ...eventBase, stage: "done" }, {
+      status: "fail",
+      message: "Max iterations reached (Solomon could not resolve)",
+      detail: failDetail
+    })
+  );
+  return { approved: false, sessionId: session.id, reason: "max_iterations" };
+}
+
+export async function tryAutoStartBoard(config, logger, emitter, eventBase) {
+  // TSK-0273: gate on hu_board.auto_start alone. The previous double-gate
+  // (enabled && auto_start) was confusing because `enabled` was never the
+  // user-facing switch — only auto_start was. This matches the AC: "kj run
+  // auto-inicia el board server si hu_board.auto_start es true".
+  if (!config.hu_board?.auto_start) return;
+  // Never auto-start during vitest: races the PID file and starts a detached
+  // process that outlives the test run.
+  if (process.env.VITEST || process.env.NODE_ENV === "test") return;
+
+  try {
+    const { startBoard, renderBoardBanner } = await import("../../commands/board.js");
+    const boardPort = config.hu_board.port || 4000;
+    const boardResult = await startBoard(boardPort);
+    const status = boardResult.alreadyRunning ? "already running" : "started";
+    // Highlighted URL box — visible regardless of log level (matches the
+    // post-planner auto-HU banner so the user gets the same signal either
+    // way the board got started).
+    console.log(renderBoardBanner({ url: boardResult.url, status }));
+    logger.info(`HU Board ${status} at ${boardResult.url}`);
+    emitProgress(emitter, makeEvent("board:started", eventBase, {
+      message: `HU Board running at ${boardResult.url}`,
+      detail: { pid: boardResult.pid, port: boardPort, url: boardResult.url }
+    }));
+  } catch (err) {
+    logger.warn(`HU Board auto-start failed (non-blocking): ${err.message}`);
+  }
+}
+
+export async function writeHistoryRecord({ sessionId, task, result, logger }) {
+  try {
+    const { createHistoryRecord } = await import("../../hu/store.js");
+    const approved = Boolean(result?.approved);
+    const summary = result?.review?.summary || result?.reason || null;
+    await createHistoryRecord(sessionId, {
+      task,
+      result: JSON.stringify(result),
+      approved,
+      summary
+    });
+  } catch (err) {
+    logger.warn(`HU history record failed (non-blocking): ${err.message}`);
+  }
+}
