+++ src/orchestrator/flow-runner.js
@@ -65,6 +65,22 @@ import {
   formatIteration, formatDecision, buildPlanSummary
 } from "./session-journal.js";
 
+// Drivers extracted from this god-module in TSK-0335 (Oleada 3 of the v2.7.4
+// audit refactor). Each driver covers one phase of the pipeline; flow-runner
+// keeps only the top-level runFlow/resumeFlow orchestration.
+import {
+  handleStandbyResult,
+  handleSolomonCheck,
+  handleReviewerRetryAndSolomon
+} from "./drivers/error-recovery.js";
+import {
+  handlePostLoopStages,
+  finalizeApprovedSession,
+  handleMaxIterationsReached,
+  tryAutoStartBoard,
+  writeHistoryRecord
+} from "./drivers/post-loop.js";
+
 // Public re-exports (loadProductContext, shouldAutoContinueCheckpoint,
 // parseCheckpointAnswer) live in src/orchestrator.js (the barrel).
 
@@ -122,423 +138,10 @@ async function runPlanningPhases({ config, logger, emitter, eventBase, session, 
 
 
 
-async function handleStandbyResult({ stageResult, session, emitter, eventBase, i, stage, logger, config, askQuestion }) {
-  if (stageResult?.action !== "standby") {
-    return { handled: false };
-  }
-
-  const isOutage = stageResult.standbyInfo.isProviderOutage;
-  const agent = stageResult.standbyInfo.agent;
-  const cooldownUntil = stageResult.standbyInfo.cooldownUntil;
-  const cooldownMs = stageResult.standbyInfo.cooldownMs;
-
-  // Rate limit = out of normal flow → Solomon decides immediately
-  const solomonResult = await invokeSolomon({
-    config, logger, emitter, eventBase, stage: `${stage}_rate_limit`, askQuestion, session,
-    iteration: i,
-    conflict: {
-      stage: `${stage}_rate_limit`,
-      task: session.task,
-      iterationCount: i,
-      maxIterations: config?.max_iterations || 5,
-      cooldownUntil,
-      cooldownMs,
-      history: [{
-        agent: stage,
-        feedback: `Agent "${agent}" rate-limited. ${isOutage ? "Provider outage (5xx)." : "API rate limit (429)."} ${cooldownUntil ? `Cooldown until ${cooldownUntil}.` : ""}`
-      }]
-    }
-  });
-
-  if (solomonResult.action === "approve") {
-    // Solomon says skip is safe — only allowed after exhausting alternatives
-    logger.info(`Solomon: skip ${stage} after evaluating risk (agent "${agent}" rate-limited)`);
-    emitProgress(emitter, makeEvent(`${stage}:rate_limit`, { ...eventBase, stage }, {
-      status: "ok",
-      message: `Solomon: skip ${stage} (low risk, agent "${agent}" unavailable)`,
-      detail: { agent, solomonAction: "approve" }
-    }));
-    return { handled: true, action: "skip" };
-  }
-
-  if (solomonResult.action === "continue") {
-    // Solomon says: wait for cooldown, or retry with alternative agent
-    const altAgent = solomonResult.alternativeAgent;
-    const waitTarget = solomonResult.waitUntil;
-
-    if (waitTarget) {
-      const waitMs = Math.max(0, new Date(waitTarget).getTime() - Date.now());
-      if (waitMs > 0 && waitMs < 10 * 60 * 1000) {
-        logger.info(`Solomon: wait ${Math.round(waitMs / 1000)}s for cooldown, then retry ${stage}`);
-        await new Promise(r => setTimeout(r, waitMs));
-      }
-    }
-
-    if (altAgent) {
-      logger.info(`Solomon: retry ${stage} with alternative agent "${altAgent}"`);
-      session._alternative_agent = { stage, provider: altAgent };
-    }
-
-    if (solomonResult.humanGuidance) {
-      session.last_reviewer_feedback = `Solomon guidance: ${solomonResult.humanGuidance}`;
-    }
-    await saveSession(session);
-    return { handled: true, action: "retry_reviewer_only" };
-  }
-
-  if (solomonResult.action === "pause") {
-    return {
-      handled: true,
-      action: "return",
-      result: { paused: true, sessionId: session.id, question: solomonResult.question, context: "rate_limit" }
-    };
-  }
-
-  // Solomon couldn't resolve — pause
-  await pauseSession(session, {
-    question: `Agent "${agent}" rate-limited. Solomon could not resolve.`,
-    context: { iteration: i, stage, reason: "rate_limit" }
-  });
-  return {
-    handled: true,
-    action: "return",
-    result: { paused: true, sessionId: session.id, question: `Agent "${agent}" rate-limited`, context: "rate_limit" }
-  };
-}
 
 
-function emitSolomonAlerts(alerts, emitter, eventBase, logger) {
-  for (const alert of alerts) {
-    emitProgress(emitter, makeEvent("brain:rules-alert", { ...eventBase, stage: "brain" }, {
-      status: alert.severity === "critical" ? "fail" : "warn",
-      message: alert.message,
-      detail: alert.detail
-    }));
-    logger.warn(`Rules alert [${alert.rule}]: ${alert.message}`);
-  }
-}
-
-async function handleSolomonCheck({ config, session, emitter, eventBase, logger, task, i, askQuestion, ciEnabled, blockingIssues, brainCtx }) {
-  if (config.pipeline?.solomon?.enabled === false) return { action: "continue" };
-
-  try {
-    const { evaluateRules, buildRulesContext } = await import("./solomon-rules.js");
-    const rulesContext = await buildRulesContext({ session, task, iteration: i, blockingIssues });
-    const rulesResult = evaluateRules(rulesContext, config.solomon?.rules);
-
-    if (rulesResult.alerts.length > 0) {
-      emitSolomonAlerts(rulesResult.alerts, emitter, eventBase, logger);
-      // Brain gateway: when Brain is the orchestrator, rule alerts are telemetry.
-      // On critical alerts Brain consults Solomon (AI judge). Only if Solomon can't
-      // resolve does Brain escalate to human. Solomon-rules never prompts directly.
-      if (!brainCtx?.enabled) {
-        const pauseResult = await checkSolomonCriticalAlerts({ rulesResult, askQuestion, session, i });
-        if (pauseResult) return pauseResult;
-      } else if (rulesResult.hasCritical) {
-        const criticalAlerts = rulesResult.alerts.filter(a => a.severity === "critical");
-        brainCtx.ruleAlerts = brainCtx.ruleAlerts || [];
-        brainCtx.ruleAlerts.push(...criticalAlerts);
-        logger.info(`Brain: ${criticalAlerts.length} critical rule alert(s) — consulting Solomon`);
-
-        const { invokeSolomon } = await import("./solomon-escalation.js");
-        const alertSummary = criticalAlerts.map(a => a.message).join("; ");
-        const solomonOpinion = await invokeSolomon({
-          config, logger, emitter, eventBase, stage: "brain-dilemma", askQuestion, session, iteration: i,
-          conflict: {
-            stage: "brain-dilemma",
-            task,
-            iterationCount: i,
-            maxIterations: config.max_iterations,
-            dilemma: `Brain detected critical rule alerts: ${alertSummary}. Should we continue iterating, pause for human, or stop?`,
-            ruleAlerts: criticalAlerts,
-            blockingIssues: blockingIssues || [],
-            history: (session.checkpoints || []).filter(cp => cp.stage === "reviewer").slice(-5).map(cp => ({ iteration: cp.iteration, feedback: cp.note || "" }))
-          }
-        });
-
-        if (solomonOpinion.action === "pause") {
-          logger.info("Brain: Solomon advised pause — escalating to human");
-          return { action: "return", result: { paused: true, sessionId: session.id, question: solomonOpinion.question || `Brain+Solomon paused: ${alertSummary}`, context: "brain_solomon_dilemma" } };
-        }
-        if (solomonOpinion.action === "approve") {
-          logger.info("Brain: Solomon advised proceeding — treating as approved");
-          return { action: "continue", approved: true };
-        }
-        // action === "continue" | "subtask" | fallback → Brain continues the loop
-        logger.info(`Brain: Solomon said '${solomonOpinion.action}' — continuing loop`);
-      }
-    }
-
-    if (ciEnabled && session.ci_pr_number) {
-      const alerts = rulesResult.alerts || [];
-      const alertMsg = alerts.length > 0
-        ? alerts.map(a => `- [${a.severity}] ${a.message}`).join("\n")
-        : "No anomalies detected";
-      await tryCiComment({
-        config, session, logger,
-        agent: "Solomon",
-        body: `Supervisor check iteration ${i}: ${alertMsg}`
-      });
-    }
-  } catch (err) {
-    logger.warn(`Solomon rules evaluation failed: ${err.message}`);
-  }
-
-  return { action: "continue" };
-}
-
-async function checkSolomonCriticalAlerts({ rulesResult, askQuestion, session, i }) {
-  if (!rulesResult.hasCritical || !askQuestion) return null;
-
-  const alertSummary = rulesResult.alerts
-    .filter(a => a.severity === "critical")
-    .map(a => a.message)
-    .join("\n");
-  const question = [
-    "Solomon detected critical issues:",
-    alertSummary,
-    "",
-    "What should we do?",
-    "1 = Continue anyway",
-    "2 = Pause the session",
-    "3 = Stop the session",
-    "",
-    "Type 1, 2, or 3:"
-  ].join("\n");
-  const answer = await askQuestion(question, { iteration: i, stage: "solomon" });
-  const trimmed = (answer || "").trim().toLowerCase();
-  const shouldPause = !answer
-    || trimmed === "2" || trimmed === "3"
-    || trimmed.startsWith("pause") || trimmed.startsWith("stop");
-  if (shouldPause) {
-    await pauseSession(session, {
-      question: `Solomon supervisor paused: ${alertSummary}`,
-      context: { iteration: i, stage: "solomon", alerts: rulesResult.alerts }
-    });
-    return { action: "pause", result: { paused: true, sessionId: session.id, reason: "solomon_alert" } };
-  }
-  return null;
-}
-
-
-async function handlePostLoopStages({ config, session, emitter, eventBase, coderRole, trackBudget, i, task, stageResults, ciEnabled, testerEnabled, securityEnabled, askQuestion, logger, brainCtx }) {
-  const postLoopDiff = await generateDiff({ baseRef: session.session_start_sha });
-
-  if (testerEnabled) {
-    const testerResult = await runTesterStage({
-      config, logger, emitter, eventBase, session, coderRole, trackBudget,
-      iteration: i, task, diff: postLoopDiff, askQuestion
-    });
-    if (testerResult.action === "pause") return { action: "return", result: testerResult.result };
-    if (testerResult.action === "continue") {
-      const summary = testerResult.stageResult?.summary || "Tester found issues";
-      session.last_reviewer_feedback = `Tester FAILED — fix these issues:\n${summary}`;
-      await saveSession(session);
-      if (testerResult.stageResult) stageResults.tester = testerResult.stageResult;
-      // Brain: push tester failure into feedback queue + compress for next coder iteration
-      if (brainCtx?.enabled) {
-        const { processRoleOutput } = await import("./brain-coordinator.js");
-        processRoleOutput(brainCtx, { roleName: "tester", output: testerResult.stageResult, iteration: i });
-      }
-      return { action: "continue" };
-    }
-    if (testerResult.stageResult) {
-      stageResults.tester = testerResult.stageResult;
-      await tryCiComment({ config, session, logger, agent: "Tester", body: `Tests: ${testerResult.stageResult.summary || "completed"}` });
-    }
-  }
-
-  if (securityEnabled) {
-    const securityResult = await runSecurityStage({
-      config, logger, emitter, eventBase, session, coderRole, trackBudget,
-      iteration: i, task, diff: postLoopDiff, askQuestion, brainCtx
-    });
-    if (securityResult.action === "pause") return { action: "return", result: securityResult.result };
-    if (securityResult.action === "continue") {
-      const summary = securityResult.stageResult?.summary || "Security found issues";
-      session.last_reviewer_feedback = `Security FAILED — fix these issues:\n${summary}`;
-      await saveSession(session);
-      if (securityResult.stageResult) stageResults.security = securityResult.stageResult;
-      // Brain: push security findings into feedback queue + compress for next coder iteration
-      if (brainCtx?.enabled) {
-        const { processRoleOutput } = await import("./brain-coordinator.js");
-        processRoleOutput(brainCtx, { roleName: "security", output: securityResult.stageResult, iteration: i });
-      }
-      return { action: "continue" };
-    }
-    if (securityResult.stageResult) {
-      stageResults.security = securityResult.stageResult;
-      await tryCiComment({ config, session, logger, agent: "Security", body: `Security scan: ${securityResult.stageResult.summary || "completed"}` });
-    }
-  }
-
-  // Final audit — last quality gate before declaring success
-  const auditResult = await runFinalAuditStage({
-    config, logger, emitter, eventBase, session, coderRole, trackBudget,
-    iteration: i, task, diff: postLoopDiff
-  });
-  if (auditResult.stageResult) {
-    stageResults.audit = auditResult.stageResult;
-    await tryCiComment({ config, session, logger, agent: "Audit", body: `Final audit: ${auditResult.stageResult.summary || "completed"}` });
-  }
-  if (auditResult.action === "retry") {
-    // Audit found actionable issues — loop back to coder
-    session.last_reviewer_feedback = auditResult.feedback;
-    await saveSession(session);
-    return { action: "continue" };
-  }
-
-  return { action: "proceed" };
-}
-
-async function finalizeApprovedSession({ config, gitCtx, task, logger, session, stageResults, emitter, eventBase, budgetSummary, pgCard, pgProject, review, i, rtkTracker }) {
-  const gitResult = await finalizeGitAutomation({ config, gitCtx, task, logger, session, stageResults });
-
-  // Accumulate final commits for PG card lifecycle tracking
-  if (gitResult?.commits?.length) {
-    const { accumulateCommit } = await import("../planning-game/pipeline-adapter.js");
-    for (const c of gitResult.commits) accumulateCommit(session, c);
-  }
-
-  if (stageResults.planner?.ok) {
-    stageResults.planner.completedSteps = [...(stageResults.planner.steps ?? [])];
-  }
-  session.budget = budgetSummary();
-  await markSessionStatus(session, "approved");
-
-  const { markPgCardToValidate } = await import("../planning-game/pipeline-adapter.js");
-  await markPgCardToValidate({ pgCard, pgProject, config, session, gitResult, logger });
-
-  const deferredIssues = session.deferred_issues || [];
-  const rtkSavings = rtkTracker?.hasData() ? rtkTracker.summary() : undefined;
-  if (rtkSavings) session.rtk_savings = rtkSavings;
-  await saveSession(session);
-
-  // --- Journal: write final files ---
-  const journalDir = session._journalDir;
-  if (journalDir) {
-    try {
-      await writeIterationsJournal(journalDir, session._journalIterations || []);
-      // decisions.md: writes only when Solomon was invoked (AC of KJC-TSK-0287).
-      // Pulls structured data from session.solomonRulings + session.brainDecisions.
-      const { writeDecisionsJournal: writeDecisionsJournalV2 } =
-        await import("../session/journal/decisions-writer.js");
-      const decisionsResult = await writeDecisionsJournalV2(session, { journalDir, logger });
-      // KJC-TSK-0288: use the new tree-writer with directory grouping
-      const { writeTreeJournal: writeTreeJournalV2 } =
-        await import("../session/journal/tree-writer.js");
-      const treeResult = await writeTreeJournalV2(journalDir, session.session_start_sha, { logger });
-
-      const journalFiles = [...(session._journalFiles || [])];
-      if (session._journalIterations?.length) journalFiles.push("iterations.md");
-      if (decisionsResult.written) journalFiles.push("decisions.md");
-      if (treeResult.written) journalFiles.push("tree.txt");
-
-      // KJC-TSK-0289: use the richer summary writer (stages table, budget
-      // breakdown, brain/solomon counts, typed links to other journal files).
-      const { writeSummaryJournal: writeSummaryJournalV2 } =
-        await import("../session/journal/summary-writer.js");
-      const startedAt = session._startedAt ? new Date(session._startedAt).toISOString() : undefined;
-      await writeSummaryJournalV2(journalDir, {
-        task: session.task,
-        result: "APPROVED",
-        sessionId: session.id,
-        iterations: i,
-        durationMs: Date.now() - (session._startedAt || Date.now()),
-        budget: budgetSummary(),
-        stages: stageResults,
-        commits: gitResult?.commits || [],
-        files: journalFiles,
-        startedAt,
-        finishedAt: new Date().toISOString(),
-        brainDecisions: Array.isArray(session.brainDecisions) ? session.brainDecisions.length : 0,
-        solomonInvocations: Array.isArray(session.solomonRulings) ? session.solomonRulings.length : 0,
-        // KJC-TSK-0327: surface skill-source decisions in the summary so the
-        // post-mortem shows which process skills (addyosmani) and stack skills
-        // (OpenSkills) actually shaped the role prompts.
-        skills: {
-          addyosmani: session.addyosmaniSkills,
-          installed: session.autoInstalledSkills,
-          recommended: session.skillsRecommended,
-        },
-      }, { logger });
-      logger.info(`Session journal written to ${journalDir}`);
-    } catch (err) {
-      logger.warn(`Journal write failed (non-blocking): ${err.message}`);
-    }
-  }
-
-  const endDetail = { approved: true, iterations: i, stages: stageResults, git: gitResult, budget: budgetSummary(), deferredIssues };
-  if (rtkSavings) endDetail.rtk_savings = rtkSavings;
-  emitProgress(
-    emitter,
-    makeEvent("session:end", { ...eventBase, stage: "done" }, {
-      message: deferredIssues.length > 0
-        ? `Session approved (${deferredIssues.length} deferred issue(s) tracked as tech debt)`
-        : "Session approved",
-      detail: endDetail
-    })
-  );
-  const result = { approved: true, sessionId: session.id, review, git: gitResult, deferredIssues };
-  if (rtkSavings) result.rtk_savings = rtkSavings;
-  return result;
-}
 
 // PG card "To Validate" logic moved to src/planning-game/pipeline-adapter.js → markPgCardToValidate()
-
-async function handleReviewerRetryAndSolomon({ config, session, emitter, eventBase, logger, review, task, i, askQuestion }) {
-  session.last_reviewer_feedback = review.blocking_issues
-    .map((x) => {
-      const parts = [`[${x.severity || "high"}] ${x.id || "ISSUE"}: ${x.description || "Missing description"}`];
-      if (x.file) parts.push(`  File: ${x.file}${x.line ? `:${x.line}` : ""}`);
-      if (x.suggested_fix) parts.push(`  Fix: ${x.suggested_fix}`);
-      return parts.join("\n");
-    })
-    .join("\n\n");
-  session.reviewer_retry_count = (session.reviewer_retry_count || 0) + 1;
-  await saveSession(session);
-
-  const maxReviewerRetries = config.session.max_reviewer_retries ?? config.session.fail_fast_repeats;
-  if (session.reviewer_retry_count < maxReviewerRetries) {
-    return { action: "continue" };
-  }
-
-  emitProgress(
-    emitter,
-    makeEvent("solomon:escalate", { ...eventBase, stage: "reviewer" }, {
-      message: `Reviewer sub-loop limit reached (${session.reviewer_retry_count}/${maxReviewerRetries})`,
-      detail: { subloop: "reviewer", retryCount: session.reviewer_retry_count, limit: maxReviewerRetries }
-    })
-  );
-
-  const solomonResult = await invokeSolomon({
-    config, logger, emitter, eventBase, stage: "reviewer", askQuestion, session, iteration: i,
-    conflict: {
-      stage: "reviewer",
-      task,
-      iterationCount: session.reviewer_retry_count,
-      maxIterations: maxReviewerRetries,
-      history: [{ agent: "reviewer", feedback: session.last_reviewer_feedback }]
-    }
-  });
-
-  if (solomonResult.action === "pause") {
-    return { action: "return", result: { paused: true, sessionId: session.id, question: solomonResult.question, context: "reviewer_fail_fast" } };
-  }
-  if (solomonResult.action === "continue") {
-    if (solomonResult.humanGuidance) {
-      session.last_reviewer_feedback += `\nUser guidance: ${solomonResult.humanGuidance}`;
-    }
-    session.reviewer_retry_count = 0;
-    await saveSession(session);
-    return { action: "continue" };
-  }
-  if (solomonResult.action === "subtask") {
-    return { action: "return", result: { paused: true, sessionId: session.id, subtask: solomonResult.subtask, context: "reviewer_subtask" } };
-  }
-
-  return { action: "continue" };
-}
 
 
 async function runPreLoopStages({ config, logger, emitter, eventBase, session, flags, pipelineFlags, coderRole, trackBudget, task, askQuestion, pgTaskId, pgProject, stageResults, brainCtx }) {
@@ -1321,160 +924,6 @@ async function handleApprovedReview({ config, session, emitter, eventBase, coder
   return { action: "return", result };
 }
 
-async function handleMaxIterationsReached({ session, budgetSummary, emitter, eventBase, config, stageResults, logger, askQuestion, task, rtkTracker, brainCtx }) {
-  const budget = budgetSummary();
-
-  // Brain-owned decision: max_iterations is guidance, not a hard rule.
-  // Brain evaluates the feedback queue state to decide extend / finalize / escalate.
-  // Solomon is only consulted if Brain cannot decide on its own.
-  if (brainCtx?.enabled) {
-    const entries = brainCtx.feedbackQueue?.entries || [];
-    const pending = entries.map(e => ({ source: e.source, category: e.category, severity: e.severity, description: e.description }));
-    const hasSecurity = entries.some(e => e.category === "security" || e.source === "security");
-    const hasCorrectness = entries.some(e => ["correctness", "tests"].includes(e.category));
-    const hasStyleOnly = entries.length > 0 && !hasSecurity && !hasCorrectness;
-
-    if (hasSecurity) {
-      // Brain: security issues unresolved → cannot finalize, escalate
-      logger.warn(`Brain: max_iterations reached with ${entries.filter(e => e.category === "security" || e.source === "security").length} unresolved security issue(s) — cannot finalize`);
-      return { paused: true, sessionId: session.id, question: "Brain: unresolved security issues at max_iterations. Review manually or extend pipeline.", context: "brain_security_block", pending };
-    }
-
-    if (hasCorrectness) {
-      // Brain: correctness/test issues pending. Cap at MAX_EXTENSIONS to avoid infinite extensions.
-      const MAX_EXTENSIONS = 2;
-      if (brainCtx.extensionCount >= MAX_EXTENSIONS) {
-        logger.warn(`Brain: ${brainCtx.extensionCount} extensions exhausted with correctness issues still pending — escalating to human`);
-        return { paused: true, sessionId: session.id, question: `Brain exhausted ${MAX_EXTENSIONS} extensions with correctness/tests still pending. Manual intervention needed.`, context: "brain_extension_cap", pending };
-      }
-      brainCtx.extensionCount += 1;
-      logger.info(`Brain: max_iterations reached with ${entries.filter(e => ["correctness", "tests"].includes(e.category)).length} correctness issue(s) pending — extending iterations (extension ${brainCtx.extensionCount}/${MAX_EXTENSIONS})`);
-      session.reviewer_retry_count = 0;
-      await saveSession(session);
-      return { approved: false, sessionId: session.id, reason: "max_iterations_extended", extraIterations: Math.ceil(config.max_iterations / 2) };
-    }
-
-    if (entries.length === 0) {
-      // Brain: no pending feedback → last reviewer approved, finalize
-      logger.info("Brain: max_iterations reached with clean feedback queue — finalizing as approved");
-      return { approved: true, sessionId: session.id, reason: "brain_approved" };
-    }
-
-    // hasStyleOnly: genuine dilemma → Brain consults Solomon
-    logger.info(`Brain: max_iterations with ${entries.length} style-only issue(s) — consulting Solomon on dilemma`);
-    const { invokeSolomon: invokeSolomonAI } = await import("./solomon-escalation.js");
-    const solomonResult = await invokeSolomonAI({
-      config, logger, emitter, eventBase, stage: "max_iterations", askQuestion, session,
-      iteration: config.max_iterations,
-      conflict: {
-        stage: "brain-max-iterations",
-        task,
-        iterationCount: config.max_iterations,
-        maxIterations: config.max_iterations,
-        budget_usd: budget?.total_cost_usd || 0,
-        dilemma: `Max iterations reached with ${entries.length} style-only issue(s) pending. Accept as-is or request more work?`,
-        pendingIssues: pending,
-        history: [{ agent: "pipeline", feedback: session.last_reviewer_feedback || "Max iterations reached" }]
-      }
-    });
-    // Brain applies Solomon's decision
-    if (solomonResult.action === "approve") {
-      logger.info("Brain: Solomon advised approve for style-only pending — finalizing");
-      return { approved: true, sessionId: session.id, reason: "brain_solomon_approved" };
-    }
-    if (solomonResult.action === "continue") {
-      return { approved: false, sessionId: session.id, reason: "max_iterations_extended", extraIterations: solomonResult.extraIterations || Math.ceil(config.max_iterations / 2) };
-    }
-    if (solomonResult.action === "pause") {
-      return { paused: true, sessionId: session.id, question: solomonResult.question, context: "brain_solomon_dilemma" };
-    }
-    // Fallback: escalate to human
-    return { paused: true, sessionId: session.id, question: `Brain+Solomon cannot resolve: ${entries.length} pending issue(s) at max_iterations`, context: "max_iterations" };
-  }
-
-  // Legacy path (Brain disabled): original Solomon-driven flow
-  const solomonResult = await invokeSolomon({
-    config, logger, emitter, eventBase, stage: "max_iterations", askQuestion, session,
-    iteration: config.max_iterations,
-    conflict: {
-      stage: "max_iterations",
-      task,
-      iterationCount: config.max_iterations,
-      maxIterations: config.max_iterations,
-      budget_usd: budget?.total_cost_usd || 0,
-      history: [{ agent: "pipeline", feedback: session.last_reviewer_feedback || "Max iterations reached without reviewer approval" }]
-    }
-  });
-
-  if (solomonResult.action === "approve") {
-    logger.info("Solomon approved coder's work at max_iterations checkpoint");
-    return { approved: true, sessionId: session.id, reason: "solomon_approved" };
-  }
-
-  if (solomonResult.action === "continue") {
-    if (solomonResult.humanGuidance) {
-      session.last_reviewer_feedback = `Solomon guidance: ${solomonResult.humanGuidance}`;
-    }
-    session.reviewer_retry_count = 0;
-    await saveSession(session);
-    const extraIterations = solomonResult.extraIterations || config.max_iterations;
-    return { approved: false, sessionId: session.id, reason: "max_iterations_extended", humanGuidance: solomonResult.humanGuidance, extraIterations };
-  }
-
-  if (solomonResult.action === "pause") {
-    return { paused: true, sessionId: session.id, question: solomonResult.question, context: "max_iterations" };
-  }
-
-  if (solomonResult.action === "subtask") {
-    return { paused: true, sessionId: session.id, subtask: solomonResult.subtask, context: "max_iterations_subtask" };
-  }
-
-  // Solomon couldn't resolve — fail
-  session.budget = budgetSummary();
-  const rtkSavings = rtkTracker?.hasData() ? rtkTracker.summary() : undefined;
-  if (rtkSavings) session.rtk_savings = rtkSavings;
-  await markSessionStatus(session, "failed");
-  const failDetail = { approved: false, reason: "max_iterations", iterations: config.max_iterations, stages: stageResults, budget: budgetSummary() };
-  if (rtkSavings) failDetail.rtk_savings = rtkSavings;
-  emitProgress(
-    emitter,
-    makeEvent("session:end", { ...eventBase, stage: "done" }, {
-      status: "fail",
-      message: "Max iterations reached (Solomon could not resolve)",
-      detail: failDetail
-    })
-  );
-  return { approved: false, sessionId: session.id, reason: "max_iterations" };
-}
-
-async function tryAutoStartBoard(config, logger, emitter, eventBase) {
-  // TSK-0273: gate on hu_board.auto_start alone. The previous double-gate
-  // (enabled && auto_start) was confusing because `enabled` was never the
-  // user-facing switch — only auto_start was. This matches the AC: "kj run
-  // auto-inicia el board server si hu_board.auto_start es true".
-  if (!config.hu_board?.auto_start) return;
-  // Never auto-start during vitest: races the PID file and starts a detached
-  // process that outlives the test run.
-  if (process.env.VITEST || process.env.NODE_ENV === "test") return;
-
-  try {
-    const { startBoard, renderBoardBanner } = await import("../commands/board.js");
-    const boardPort = config.hu_board.port || 4000;
-    const boardResult = await startBoard(boardPort);
-    const status = boardResult.alreadyRunning ? "already running" : "started";
-    // Highlighted URL box — visible regardless of log level (matches the
-    // post-planner auto-HU banner so the user gets the same signal either
-    // way the board got started).
-    console.log(renderBoardBanner({ url: boardResult.url, status }));
-    logger.info(`HU Board ${status} at ${boardResult.url}`);
-    emitProgress(emitter, makeEvent("board:started", eventBase, {
-      message: `HU Board running at ${boardResult.url}`,
-      detail: { pid: boardResult.pid, port: boardPort, url: boardResult.url }
-    }));
-  } catch (err) {
-    logger.warn(`HU Board auto-start failed (non-blocking): ${err.message}`);
-  }
-}
 
 async function initFlowContext({ task, config, logger, emitter, askQuestion, pgTaskId, pgProject, flags }) {
   // Auto-init .karajan/ if missing (copies coder-rules, review-rules, role templates)
@@ -1769,21 +1218,6 @@ async function runSingleIteration(ctx) {
   return { action: "next" };
 }
 
-async function writeHistoryRecord({ sessionId, task, result, logger }) {
-  try {
-    const { createHistoryRecord } = await import("../hu/store.js");
-    const approved = Boolean(result?.approved);
-    const summary = result?.review?.summary || result?.reason || null;
-    await createHistoryRecord(sessionId, {
-      task,
-      result: JSON.stringify(result),
-      approved,
-      summary
-    });
-  } catch (err) {
-    logger.warn(`HU history record failed (non-blocking): ${err.message}`);
-  }
-}
 
 /**
  * Run the standard iteration loop for a given task using the pipeline context.
